Add medication management features with database integration and alarm functionality
