+++ app/src/main/java/dev/indra/smartmedicinereminder/receiver/NotificationDismissReceiver.java
@@ -1,0 +1,14 @@
+package dev.indra.smartmedicinereminder;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+public class NotificationDismissReceiver extends BroadcastReceiver {
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        int notificationId = intent.getIntExtra("notification_id", 0);
+
+        ReminderReceiver.stopAlerts(context, notificationId);
+    }
+}
