+++ app/src/main/java/dev/indra/smartmedicinereminder/utils/AlarmHelper.java
@@ -1,0 +1,66 @@
+package dev.indra.smartmedicinereminder;
+
+import android.annotation.SuppressLint;
+import android.app.AlarmManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+
+import java.util.Calendar;
+
+public class AlarmHelper {
+
+    @SuppressLint("ScheduleExactAlarm")
+    public static void setAlarm(Context context, Medication medication) {
+        AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+        Intent intent = new Intent(context, ReminderReceiver.class);
+        intent.putExtra("medication_id", medication.getId());
+        intent.putExtra("medication_name", medication.getName());
+        intent.putExtra("medication_dosage", medication.getDosage());
+
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(
+                context,
+                (int) medication.getId(),
+                intent,
+                PendingIntent.FLAG_UPDATE_CURRENT | PendingIntent.FLAG_IMMUTABLE
+        );
+
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.HOUR_OF_DAY, medication.getHour());
+        calendar.set(Calendar.MINUTE, medication.getMinute());
+        calendar.set(Calendar.SECOND, 0);
+
+        if (calendar.getTimeInMillis() < System.currentTimeMillis()) {
+            calendar.add(Calendar.DAY_OF_MONTH, 1);
+        }
+
+        if (alarmManager != null) {
+            alarmManager.setExactAndAllowWhileIdle(
+                    AlarmManager.RTC_WAKEUP,
+                    calendar.getTimeInMillis(),
+                    pendingIntent
+            );
+        }
+    }
+
+
+    public static void updateAlarm(Context context, Medication medication) {
+        cancelAlarm(context, (int) medication.getId());
+        setAlarm(context, medication);
+    }
+
+    public static void cancelAlarm(Context context, int medicationId) {
+        AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
+        Intent intent = new Intent(context, ReminderReceiver.class);
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(
+                context,
+                medicationId,
+                intent,
+                PendingIntent.FLAG_IMMUTABLE
+        );
+
+        if (alarmManager != null) {
+            alarmManager.cancel(pendingIntent);
+        }
+    }
+}
